extract(parser result): Extract shared parser result helpers
- Move shared trivia and child-boundary helpers into dedicated utility files so parser result normalization code stays focused on grammar-specific rewrites.
- Keep behavior unchanged while centralizing reusable node logic and making the helper set easier to discover and maintain.
