+++ parser_result_trivia_helpers.go
@@ -1,0 +1,48 @@
+package gotreesitter
+
+func bytesAreTrivia(b []byte) bool {
+	for _, c := range b {
+		switch c {
+		case ' ', '\t', '\n', '\r':
+			continue
+		default:
+			return false
+		}
+	}
+	return true
+}
+
+func lastNonTriviaByteEnd(source []byte) uint32 {
+	for i := len(source); i > 0; i-- {
+		switch source[i-1] {
+		case ' ', '\t', '\n', '\r', '\f':
+			continue
+		default:
+			return uint32(i)
+		}
+	}
+	return 0
+}
+
+func trimTrailingExtraTriviaRoot(root *Node, source []byte) {
+	if root == nil || len(root.children) == 0 || len(source) == 0 {
+		return
+	}
+	last := root.children[len(root.children)-1]
+	if last == nil || !last.IsExtra() || len(last.children) != 0 {
+		return
+	}
+	if last.startByte >= last.endByte || last.endByte != root.endByte || int(last.endByte) > len(source) {
+		return
+	}
+	if !bytesAreTrivia(source[last.startByte:last.endByte]) {
+		return
+	}
+	root.children = root.children[:len(root.children)-1]
+	if len(root.fieldIDs) > len(root.children) {
+		root.fieldIDs = root.fieldIDs[:len(root.children)]
+	}
+	if len(root.fieldSources) > len(root.children) {
+		root.fieldSources = root.fieldSources[:len(root.children)]
+	}
+}
